afroken_llm_backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
asyncpg==0.27.0
psycopg2-binary==2.9.7
pgvector==0.5.0
pydantic==1.10.12
python-dotenv==1.0.0
httpx==0.24.1
redis==4.6.0
celery==5.3.1
minio==7.1.11
aiofiles==23.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.3.2
prometheus-client==0.16.0


